% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConfig.R
\name{checkConfig}
\alias{checkConfig}
\title{Checks if shinycell config data.table contains any errors}
\usage{
checkConfig(scConf, obj)
}
\arguments{
\item{scConf}{shinycell config data.table}

\item{obj}{input single-cell object for Seurat (v3+) / SingleCellExperiment 
data or input file path for h5ad / loom files}
}
\value{
any potential error messages
}
\description{
Checks if shinycell config data.table contains any errors. It is useful and 
reccomended to run this function if users have motified the shinycell 
config manually. Errors can include (i) levels in scConf does not match 
that in the Seurat/SingleCellExperiment object, (ii) number of levels does 
not match number of colours and (iii) specified colours are invalid colours.
}
\examples{
# Load example Seurat object
seu <- readRDS(system.file("extdata/readySeu_rset.rds", package = "pyShinyCell"))

# Create minimal config
scConf <- data.table::data.table(
  ID = c("orig.ident"),
  UI = c("Original Identity"),
  fID = c("H1|H2|H3|H4"),
  fUI = c("H1|H2|H3|H4"),
  default = c(1)
)

# Check configuration
checkConfig(scConf, seu)

}
\author{
John F. Ouyang
}
