% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_data.R
\name{.createCustomPalettes}
\alias{.createCustomPalettes}
\title{Custom Color Palette List}
\usage{
.createCustomPalettes()
}
\description{
A list of manually curated color palettes used as defaults in pyShinyCell.
}
\keyword{internal}
