% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_data.R
\name{.createPaletteDataFile}
\alias{.createPaletteDataFile}
\title{Create Palette Data File}
\usage{
.createPaletteDataFile()
}
\description{
Internal function to create and save the pal.info data object to
data/pal.info.rda. This is called during package build.
}
\keyword{internal}
