% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_codegen.R
\name{.gen_tab_b1_server}
\alias{.gen_tab_b1_server}
\title{Generate Tab-Specific Server Code}
\usage{
.gen_tab_b1_server(prefix)
}
\description{
Generate Tab-Specific Server Code
}
\keyword{Functions}
\keyword{analysis}
\keyword{by}
\keyword{code}
\keyword{each}
\keyword{extracting}
\keyword{for}
\keyword{from}
\keyword{generate}
\keyword{implementation.}
\keyword{internal}
\keyword{patterns}
\keyword{reference}
\keyword{server/UI}
\keyword{shinyApp_stable}
\keyword{tab}
\keyword{to}
