% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePyShinyCode.R
\name{.generateGlobalCode}
\alias{.generateGlobalCode}
\title{Generate global.R with Python setup}
\usage{
.generateGlobalCode(prefix, py_env)
}
\description{
Generate global.R with Python setup
}
\keyword{internal}
