% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePyShinyCode.R
\name{.generateServerCode}
\alias{.generateServerCode}
\title{Generate server.R code}
\usage{
.generateServerCode(scConf, prefix, enable_tabs, py_env, verbose = FALSE)
}
\description{
Generate server.R code
}
\keyword{internal}
