% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePyShinyCode.R
\name{.generateUICode}
\alias{.generateUICode}
\title{Generate ui.R code}
\usage{
.generateUICode(scConf, prefix, title, enable_tabs, verbose = FALSE)
}
\description{
Generate ui.R code
}
\keyword{internal}
