% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modColours.R
\name{modColours}
\alias{modColours}
\title{Modify the colour palette for categorical metadata}
\usage{
modColours(scConf, meta.to.mod, new.colours)
}
\arguments{
\item{scConf}{shinycell config data.table}

\item{meta.to.mod}{metadata for which to modify the colour palette. Users 
can either use the actual metadata column names or display names. Please 
specify only one metadata}

\item{new.colours}{character vector of new colour palette}
}
\value{
updated shinycell config data.table
}
\description{
Modify the colour palette for categorical metadata.
}
\examples{
# Create minimal example config with categorical metadata
scConf <- data.table::data.table(
  ID = c("library"),
  UI = c("Library"),
  fID = c("S1|S2|S3|S4"),
  fUI = c("S1|S2|S3|S4"),
  fCL = c("red|blue|green|orange"),
  default = c(0)
)
# Modify color palette
scConf <- modColours(scConf, meta.to.mod = "Library",
                     new.colours = c("black", "darkorange", "blue", "red"))

}
\author{
John F. Ouyang
}
