% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modDefault.R
\name{modDefault}
\alias{modDefault}
\title{Set the default metadata to display}
\usage{
modDefault(scConf, default1, default2)
}
\arguments{
\item{scConf}{shinycell config data.table}

\item{default1}{metadata to set as the 1st default metadata to display. 
Users can either use the actual metadata column names or display names}

\item{default2}{metadata to set as the 2nd default metadata to display. 
Users can either use the actual metadata column names or display names}
}
\value{
updated shinycell config data.table
}
\description{
Set the default metadata to display in the shiny app. Default1 is used when 
plotting metadata with gene expression or when plotting two metadata 
simultaneously. Default2 is used when plotting two metadata simultaneously.
}
\examples{
# Create minimal example config with multiple metadata
scConf <- data.table::data.table(
  ID = c("orig.ident", "seurat_clusters"),
  UI = c("Library", "Cluster"),
  fID = c("S1|S2", "C1|C2|C3"),
  fUI = c("S1|S2", "C1|C2|C3"),
  default = c(0, 0)
)
# Set default metadata to display
scConf <- modDefault(scConf, default1 = "Library", default2 = "Cluster")

}
\author{
John F. Ouyang
}
