% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLabels.R
\name{modLabels}
\alias{modLabels}
\title{Modify the legend labels for categorical metadata}
\usage{
modLabels(scConf, meta.to.mod, new.labels)
}
\arguments{
\item{scConf}{shinycell config data.table}

\item{meta.to.mod}{metadata for which to modify the legend labels. Users 
can either use the actual metadata column names or display names. Please 
specify only one metadata}

\item{new.labels}{character vector of new legend labels}
}
\value{
updated shinycell config data.table
}
\description{
Modify the legend labels for categorical metadata.
}
\examples{
# Create minimal example config with categorical metadata
scConf <- data.table::data.table(
  ID = c("library"),
  UI = c("Library"),
  fID = c("S1|S2|S3|S4"),
  fUI = c("Sample1|Sample2|Sample3|Sample4"),
  default = c(0)
)
# Modify legend labels
scConf <- modLabels(scConf, meta.to.mod = "library",
                    new.labels = c("Fib", "Primed", "Naive", "RSeT"))

}
\author{
John F. Ouyang
}
