% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_data.R
\docType{data}
\name{pal.info}
\alias{pal.info}
\title{Palette Information Data Frame}
\format{
A data frame with palette metadata.
}
\description{
A data frame containing metadata about available color palettes for use in
pyShinyCell applications. Combines custom palettes, ggsci (scientific journal)
palettes, and RColorBrewer palettes.
}
\details{
Columns:
- maxcolors: Integer. Maximum number of colors in the palette.
- category: Character. Type of palette ("qual", "seq", "div" for qualitative,
  sequential, or diverging).
- palette: Character. Palette name or NA for internal references.
- package: Character. Package providing the palette ("custom", "ggsci", or
  "RColorBrewer").
- pal.setup: List. Color vector for each palette.


This data frame is used internally by generated Shiny apps to provide
users with palette selection options. It includes:
- 6 custom palettes (sequential and diverging)
- 40 ggsci palettes (scientific journal color schemes)
- 35 RColorBrewer palettes (sequential, diverging, qualitative)
}
\examples{
# View available palettes
head(pal.info)
unique(pal.info$package)

}
\keyword{data}
