% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderMeta.R
\name{reorderMeta}
\alias{reorderMeta}
\title{Reorder the order in which metadata appear in the shiny app}
\usage{
reorderMeta(scConf, new.meta.order)
}
\arguments{
\item{scConf}{shinycell config data.table}

\item{new.meta.order}{character vector containing new order. All metadata 
names must be included, which can be found at \code{scConf$ID}}
}
\value{
updated shinycell config data.table
}
\description{
Reorder the order in which metadata appear in the dropdown menu in the 
shiny app.
}
\examples{
# Create minimal example config
scConf <- data.table::data.table(
  ID = c("cell_type", "batch", "condition"),
  UI = c("Cell Type", "Batch", "Condition"),
  fID = c("TypeA|TypeB", "B1|B2", NA),
  default = c(1, 0, 0)
)
# Reorder: batch first, then cell_type, then condition
scConf <- reorderMeta(scConf, c("batch", "cell_type", "condition"))

}
\author{
John F. Ouyang
}
