% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showLegend.R
\name{showLegend}
\alias{showLegend}
\title{Shows the legends for single-cell metadata}
\usage{
showLegend(scConf, fontSize = 14)
}
\arguments{
\item{scConf}{shinycell config data.table}

\item{fontSize}{font size of legends. Decrease it if you have too many 
items to display}
}
\value{
gtable plot showing the legends for different metadata
}
\description{
Shows the legends for single-cell metadata based on the shinycell config 
data.table. This allows user to visualise the different metadata to be 
plotted and make any modifications if necessary. Note that the display name 
is shown here instead of the actual name. For more information regarding 
display name, see \code{?modMetaName}.
}
\examples{
# Create minimal example config
scConf <- data.table::data.table(
  ID = c("orig.ident", "seurat_clusters"),
  UI = c("Library", "Cluster"),
  fID = c("S1|S2", "C1|C2|C3"),
  fUI = c("S1|S2", "C1|C2|C3"),
  fCL = c("black|red", "blue|green|orange"),
  default = c(1, 0)
)
# Show legends for the configuration
scLegend <- showLegend(scConf)

}
\author{
John F. Ouyang
}
